Fixed the logic issue of calculating the output of T-Net
Before the output depended on the dimension of the input
points, but PointNet treats as a point dimension different
axis. e.g. for input T-Net the point dimension axis = 2,
but for features the point dimenstion axis = 3. The value
of dimension is controled by value k.
Fixed implementation follows this logic now.